insurance_site/requirements.txt
apiclient==1.0.3
appdirs==1.4.0
argcomplete==1.8.2
awscli==1.11.47
backports.shutil-get-terminal-size==1.0.0
base58==0.2.4
boto3==1.4.4
botocore==1.5.26
click==6.7
colorama==0.3.7
decorator==4.0.11
discovery==0.0.4
Django==1.10.5
django-storages==1.5.2
djangorestframework==3.5.4
docutils==0.13.1
durationpy==0.3
enum34==1.1.6
futures==3.0.5
google-api-python-client==1.6.2
hjson==2.0.2
httplib2==0.10.3
ipython==5.2.2
ipython-genutils==0.1.0
jmespath==0.9.2
kappa==0.6.0
lambda-packages==0.14.2
MySQL-python==1.2.5
oauth2client==4.0.0
packaging==16.8
pathlib2==2.2.1
pickleshare==0.7.4
placebo==0.8.1
prompt-toolkit==1.0.13
pyasn1==0.2.2
pyasn1-modules==0.0.8
Pygments==2.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
python-slugify==1.2.1
PyYAML==3.12
requests==2.13.0
rsa==3.4.2
s3transfer==0.1.10
simplegeneric==0.8.1
six==1.10.0
toml==0.9.2
tqdm==4.11.2
traitlets==4.3.1
troposphere==1.9.2
Unidecode==0.4.20
uritemplate==3.0.0
urllib3==1.20
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.12
win-unicode-console==0.5
wsgi-request-logger==0.4.6
zappa==0.39.1
